backend/requirements.txt
Flask==2.2.3
Flask-Login==0.6.2
python-dotenv==1.0.0
web3==6.0.0
requests==2.28.2
Werkzeug==2.2.3
py-solc-x==1.1.1